+++ app/Http/Controllers/CountryController.php
@@ -7,17 +7,21 @@ use Illuminate\Http\Request;
 
 class CountryController extends Controller
 {
-    public function index(Country $country){
-        $org_id=orgId();
-        $countries=Country::where('organization_id',$org_id)->latest()->get();
-        return view('place.country',compact('country','countries'));
+    public function index(Country $country)
+    {
+        $org_id = orgId();
+        if (!$org_id) {
+            return redirect()->back()->with('error', "Please select an organization before perform any operation on it.");
+        }
+        $countries = Country::where('organization_id', $org_id)->latest()->get();
+        return view('place.country', compact('country', 'countries'));
     }
 
     public function store(Request $request)
     {
         $org_id = orgId();
         if (!$org_id) {
-            return redirect()->back()->with('error', "Please select an organization.");
+            return redirect()->back()->with('error', "Please select an organization before perform any operation on it.");
         }
         $data = $request->validate([
             'name' => 'required',
@@ -32,14 +36,13 @@ class CountryController extends Controller
     {
         $orgId = orgId();
         if (!$orgId) {
-            return redirect()->back()->with('error', "Please select an organization.");
+            return redirect()->back()->with('error', "Please select an organization before perform any operation on it.");
         }
         $country = Country::where('id', $id)->where('organization_id', $orgId)->first();
 
         if ($country) {
             $countries = Country::where('organization_id', $orgId)->latest()->get();
-            return view('place.country',compact('country','countries'));
-
+            return view('place.country', compact('country', 'countries'));
         }
         return redirect()->back()->with('error', "No data found");
     }
@@ -55,12 +58,16 @@ class CountryController extends Controller
 
     public function delete($id)
     {
-        $data = Country::where('id', $id)->where('organization_id', orgId())->first();
+        $org_id = orgId();
+        if (!$org_id) {
+            return redirect()->back()->with('error', "Please select an organization before perform any operation on it.");
+        }
+        $data = Country::where('id', $id)->where('organization_id', $org_id)->first();
 
-        if(!$data){
+        if (!$data) {
             return redirect()->back()->with('error', "No data found");
         }
         $data->delete();
-        return redirect()->route('country.index')->with('success',"Brand removed");
+        return redirect()->route('country.index')->with('success', "Brand removed");
     }
 }
